requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
phonenumberslite==8.12.57
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
